OmniShadowPass cubemap successful.

+++ Header/Lights/DirectionalLight.h
@@ -10,7 +10,7 @@ public:
 
     DirectionalLight(GLfloat red, GLfloat green, GLfloat blue, 
                     GLfloat _ambientIntensity, GLfloat _diffuseIntensity,
-                    GLfloat shadowWidth, GLfloat shadowHeight, GLfloat near, GLfloat far,
+                    GLuint shadowWidth, GLuint shadowHeight, GLfloat near, GLfloat far,
                     GLfloat xDir, GLfloat yDir, GLfloat zdir);
 
     void UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
