+++ Header/Lights/Light.h
@@ -1,6 +1,8 @@
 #ifndef LIGHT_H
 #define LIGHT_H
 
+#include <vector>
+
 #include <GL/glew.h>
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
@@ -13,7 +15,7 @@ public:
 
     Light();
     Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat _ambientIntensity, GLfloat _diffuseIntensity, 
-            GLfloat shadowWidth, GLfloat shadowHeight, GLfloat near, GLfloat far);
+            GLuint shadowWidth, GLuint shadowHeight, GLfloat near, GLfloat far);
 
     virtual void UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation, 
                 GLuint diffuseIntensityLocation, GLuint directionLocation=0,
