+++ Source/Lights/DirectionalLight.cpp
@@ -7,7 +7,7 @@ DirectionalLight::DirectionalLight() : Light(){
 }
 
 DirectionalLight::DirectionalLight(GLfloat red, GLfloat green, GLfloat blue, GLfloat _ambientIntensity, GLfloat _diffuseIntensity,
-                                    GLfloat shadowWidth, GLfloat shadowHeight, GLfloat near, GLfloat far,
+                                    GLuint shadowWidth, GLuint shadowHeight, GLfloat near, GLfloat far,
                                     GLfloat xDir, GLfloat yDir, GLfloat zDir) 
                                     : Light(red, green, blue, _ambientIntensity, _diffuseIntensity, shadowWidth, shadowHeight, near, far){
